samod: Switch to doc_auto_cfg
Problem: we were using `cfg_attr(docrs, doc(cfg(...)))` in many places
to generate nice "only available if" messages in the docs. This leads to
lots of repetition and is easy to forget to add in new places.

Solution: use the new `doc_auto_cfg` feature of rustdoc[1] which does
the same thing automatically.

[1]: https://doc.rust-lang.org/beta/unstable-book/language-features/doc-auto-cfg.html#doc_auto_cfg