Switch from futures::channel::mpsc and std::sync::mpsc::channel to async_channel
Currently document actors run on a `rayon` thread pool and use a
blocking `std::sync::mpsc::channel` to communicate with the main thread.
I want to make the use of a thread pool optional, spawning the actors
onto the runtime if there is no thread pool configured. This means that
the communication channel between the main thread and the actors must be
capable of async operations. `async_channel` is a stable, and widely
used async channel implementation that works in both sync and async
contexts. For consistency we use it everywhere, replacing also the uses
of `futures::channel::mpsc`.