fix: don't delete the current compacted chunk when compacting
Problem: it is possible for the compaction logic to lose data if the
following sequence of events occurs:

1. Enough chunks are created to trigger a compaction. Storage looks
   like this:
    - /<document ID>/incremental/chunk1 
    ...
    - /<document ID>/incremental/chunk100 
2. The compaction process writes a new compacted chunk. Storage like
   this:
    - /<document ID>/incremental/chunk1 
    ...
    - /<document ID>/incremental/chunk100 
    - /<document ID>/snapshot/compacted_chunk1
3. Before the compaction process can delete the old chunks, the process
   is killed for some reason
4. The process restarts, and the compaction process runs again, seeing
   the same number of chunks as before. This triggers compaction. 
5. Once the new compacted chunk is written, the old chunks are deleted,
   but "old chunks" now includes the compacted chunk because it's the
   same as the chunk written in step 2.

The problem is that when the compaction process runs, it deletes
everything that it read from disk, even if this includes the compacted
chunk it is currently writing.

Solution: when deleting old chunks, don't delete the chunk matching the
one currently being written.

While we're here I also made the `CompactionHash` struct public, as it's
useful for testing and might be useful elsewhere as well.