samod-core: don't request from everyone whilst syncing
The existing logic in `samod-core` when we receive a sync message or
request for a document we don't currently have is to immediately send a
request to every other peer we are connected to. This causes problems if
the incoming message is a sync message.

Say we have three peers connected like this:

alice <-> bob <-> charlie <-> derek

Alice is configured to announce everything to bob. Say Alice creates a
document and Derek queries for it before the sync from Alice to Bob to
Charlie has completed. The request handling logic will result in the
following sequence of events:

* Alice announces document to bob
* Bob begins syncing with Alice
* Bob also sends a request to Charlie for the document
* Derek begins finding the document
* Derek sends a request to Charlie

At this point Charlie has received a request from Bob and a request from
Derek, from Charlies perspective no-one has the document so he returns a
not-available response to Derek.

To solve this we don't send outgoing request messages whilst we have at
least one peer who is syncing with us.