samod: make threadpool usage compile time optional
Problem: the use of the rayon threadpool option to run document actors
in paralell requires the use of the
`async_channel::Sender::blocking_send` and
`async_channel::Receiver::recv_blocking` methods, which are behind a
conditional compilation flag which is not enabled for `wasm` targets.
This means that `samod` cannot be used on `wasm` targets even if the
threadpool feature is not being used.

Solution: make the threadpool feature compile time optional behind the
`threadpool` feature flag. We also put the `blocking_send` and
`recv_blocking` methods behind the same flag so that if the `threadpool`
feature is not available `samod` can compile on `wasm` just fine. While
we're here add a newtype wrapper for unbounded channels to make it
clearer that we expect these channels to be unbounded.