Add explicit Connection abstraction
Problem: it is often necessary to wait for a connection to be running
before taking action, or to examine a connection to find out the peer ID
of the other end. Furthermore, driving connection futures returned by
`Repo::connect` is awkward and easy to forget.

Solution: Introduce a `Connection` abstraction which is immediately
(synchronously) returned by `Repo::connect` and represents the ongoing
state of the connection. This gives us a place to put APIs for
connection related actions such as waiting for the handshake to
complete. To achieve this we run the connection handling logic on the IO
loop, which removes the need for the user to drive a future returned by
`Repo::connect`.